// src/app/modules/models/stampingEvent.model.ts
export class StampingEvent{
    id: number;
    idDip: number;
    desDip: string;
    start: string;
    end: string;
    ordinary:  number;
    absence: number;
    surplus:  number;
    type:string;
    day:  number;
    month:  number;
    year: number;
    note:string;
    other: any[];
    just: any[];
    idTask: number;
    idProject: number;
    status: boolean;
}