// src/app/modules/models/fitokData.model.ts
export class FitokData{
  id: number;
  fitok: string;
  dateStartProduction:  number;
  dateEndProduction:  number;
  dateDelivery:  number;
  sellingPrice: number;
  ddt:  string;
  type:  string;
  material:  string;
  length: number;
  width: number;
  height: number;
  assembly:  string;
  tare: number;
  net: number;
  gross: number;
  quantity: number;
  firCm: number;
  plywoodCm: number;
  osbCm: number;
  totalCm: number;
}